+++ t/1144-isf-public-tested-by-metadata-audit.t
@@ -238,6 +238,7 @@ my $expected_tested_by = [
     't/1344-isf-assemble-static-part-widths.t',
     't/1345-isf-actor-param-actor-constants.t',
     't/1346-isf-actor-param-actor-params.t',
+    't/1347-isf-transaction-param-actor-static-defaults.t',
 ];
 
 subtest 'direct ISF tested_by metadata is exact and valid' => sub {